Put a separate Page Template component
