+++ Showroom/Main/index.jsx
@@ -1,15 +1,20 @@
 import React from 'react'
 import Grid from '../Grid'
-import Page from '../Page'
+import * as Page from '../Page'
 import Sidebar from '../Sidebar'
 import GettingStarted from '../Page/GettingStarted'
+import { has } from 'ramda'
 
 import 'normalize.css'
 import './index.scss'
 
 const getCurrentPage = ([anchor], examples) =>
   anchor
-    ? <Page example={examples.components[anchor] || examples.templates[anchor]} />
+    ? (
+      has(anchor, examples.components)
+        ? <Page.Component example={examples.components[anchor]} />
+        : <Page.Template example={examples.templates[anchor]} />
+    )
     : <GettingStarted />
 
 export default function Showroom ({route, grid, examples}) {
